+++ CHANGES.md
@@ -11,6 +11,7 @@
   `adw::Dialog`
 + macros: Allow setting conditional widget `gtk::Stack` properties
 + core: Implement factory view for `gtk::Fixed`
++ macros: Conditional root widgets
 
 ### Changed